+++ src/index.ts
@@ -1,2 +1,3 @@
 export { Seq } from "./Seq";
 export * from "./static";
+export * from "./combinatorics";